Include latest shrink and stretch class
